Separate GlowHoverContainer from BaseContainer
