+++ ui/src/components/containers/index.ts
@@ -1,1 +1,2 @@
 export * from './base-container';
+export * from './glow-hover-container';
