+++ website/src/app/[lang]/[region]/(website)/(home)/(sections)/approach.tsx
@@ -1,6 +1,7 @@
 import { DefaultParams } from '@/app/[lang]/[region]';
 import { Translator } from '@socialincome/shared/src/utils/i18n';
 import { Card, CardContent, CardHeader, Typography } from '@socialincome/ui';
+import { useGlowHover } from '@socialincome/ui';
 
 type ApproachCardProps = {
 	category: string;
@@ -24,7 +25,7 @@ export async function Approach({ lang }: DefaultParams) {
 			</Typography>
 			<div className="mx-auto flex w-full max-w-7xl flex-wrap gap-2">
 				{cards.map((card) => (
-					<Card key={card.title} className="theme-blue mx-auto min-w-80 max-w-md flex-1 rounded-none border-none">
+					<Card hasGlowHover key={card.title} className="theme-blue mx-auto min-w-80 max-w-md flex-1 rounded-none border-none">
 						<CardHeader className="mb-1">
 							<Typography size="md" className="opacity-40">
 								{card.category}
